requirements.txt
pyinstaller==5.13.0
google-api-python-client
viam-sdk
